cue: default to int64 in Value.Decode for integer values
We were defaulting to int rather than int32 or int64, so the behavior
changed depending on whether the host platform was 32-bit or 64-bit.

This got caught by the encoding/toml tests, which use a hexadecimal
value which happens to be large enough to not fit into int32.
Thus, the tests roundtripping from TOML to CUE and back to TOML would
make the value overflow and become negative, causing a test failure.

This issue was not covered in TestDecode as none of the test cases
used large enough numbers. Add them, verifying that they showed
the wrong behavior on GOARCH=386.

    We can soon start testing this by adding a CI step on amd64 like:

        GOARCH=386 go test -short ./...

Updates #3540.

+++ mod/modfile/modfile_test.go
@@ -404,7 +404,7 @@ other: field: 123
 		Custom: map[string]map[string]any{
 			"legacy": {
 				"some":  true,
-				"other": map[string]any{"field": 123},
+				"other": map[string]any{"field": int64(123)},
 			},
 		},
 	},
